requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
PyMuPDF==1.26.4
sentence-transformers==5.1.0
chromadb==1.0.20
google-generativeai==0.8.5
python-multipart==0.0.20